chore: refactor exec.rs: create separate seatbelt.rs and spawn.rs files
